+++ src/Voucherify.Client.net45.Tests/ApiCustomers.cs
@@ -1,0 +1,10 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Voucherify.Client.net45.Tests
+{
+    [TestClass]
+    public class ApiCustomers
+    {
+
+    }
+}