+++ src/Voucherify.Client.net45.Tests/ApiRedemptions.cs
@@ -1,0 +1,9 @@
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
+
+namespace Voucherify.Client.net45.Tests
+{
+    [TestClass]
+    public class ApiRedemptions
+    {
+    }
+}