+++ src/Voucherify.Client.net45.Tests/Configuration.cs
@@ -1,0 +1,8 @@
+﻿namespace Voucherify.Client.net45.Tests
+{
+    public static class Configuration
+    {
+        public const string AppId = "<your_appid>";
+        public const string AppToken = "<your_app_token>";
+    }
+}