+++ src/Voucherify.Client/DataModel/Metadata.cs
@@ -6,5 +6,12 @@ namespace Voucherify.Client.DataModel
     [JsonObject]
     public class Metadata : Dictionary<string, object>
     {
+        public Metadata()
+        {
+
+        }
+        public Metadata(Dictionary<string, object> dictionary) : base(dictionary)
+        {
+        }
     }
 }